跑起来了

+++ RP/UtilShader.h
@@ -8,7 +8,7 @@ private:
 	Camera* camera;
 	Vec3f lightDir;
 public:
-	Vec3i  vertex(Vec3f worldPos, Vec3f normal, Vec2f uv, int nvert) override;
+	Vec3f  vertex(Vec3f worldPos, Vec3f normal, Vec2f uv, int nvert) override;
 	Color  fragment(Vec3f bary) override;
 	UtilShader(Camera* camera, Vec3f lightDir) :camera(camera), lightDir(lightDir),
 	diffuse(255,255,255){}
